+++ assets/webpack/src/js/chart.js
@@ -4,15 +4,26 @@ import Tooltip from 'bootstrap/js/dist/tooltip';
 import * as storage from './storage';
 import * as helpers from './helpers';
 
-// Charts are self-conscious about their visibility to avoid being rendered
-// if they are not visible to the user. To accomplish this, a global
-// 'IntersectionObserver' is used to monitor the visibility of the charts and
-// update their status accordingly.
-const observer = new IntersectionObserver((entries) => {
+// Charts are aware of their visibility to avoid rendering when not visible and
+// to start/stop refresh loops accordingly. A global 'IntersectionObserver' is
+// used to monitor their visibility status.
+const intersectionObserver = new IntersectionObserver((entries) => {
   entries.forEach(entry => {
-    entry.target.chart.setVisible(entry.isIntersecting);
+    entry.target.chart.handleVisibilityChange(entry.isIntersecting);
   });
 }, { threshold: 0.1 });
+
+// Charts need to be notified about changes in their container size to adjust
+// the graph accordingly. This might require a full data refresh (e.g., less
+// space is available, resulting in a smaller effective step, which requires
+// fetching new down-sampled data), or just a re-render of the graph (e.g., more
+// space is available, allowing the use of lines and markers instead of just
+// lines).
+const resizeObserver = new ResizeObserver((entries) => {
+  entries.forEach(entry => {
+    entry.target.chart.handleSizeChange();
+  });
+});
 
 class Chart {
   constructor(container, metric, rangeFactory, refreshInterval, aggregator, step) {
@@ -25,38 +36,87 @@ class Chart {
 
     this.listeners = {};
 
+    this.initializing = false;
+    this.refreshing = false;
+
+    this.debouncedHandleRefresh = helpers.debounce(this.handleRefresh.bind(this), 500);
+
     this.visible = false;
-    this.graph = null;
     this.interval = null;
+    this.lastRefresh = 0;
+    this.pendingRefresh = false;
     this.error = null;
 
-    this.range = null;
-    this.zoomRange = null;
+    this.graph = {
+      // A lazily initialized reference to the Plotly graph element.
+      element: null,
 
-    observer.observe(this.container);
+      // The data currently displayed in the graph, as returned by the storage,
+      // but slightly adjusted: hex bitmaps as integers, etc. Currently,
+      // keeping this data around is not strictly necessary, but it might be
+      // useful in the future to support incremental updates.
+      x: null,
+      y: null,
+
+      // The step of the data currently displayed in the graph, as returned by
+      // the storage. This is currently needed to estimate the optimal mode for
+      // the Plotly data depending on the available space (i.e., number of
+      // columns, zoom level, etc.).
+      step: null,
+
+      // The range in the X axis for the data currently contained in the graph,
+      // as returned by the storage, but slightly adjusted: right boundary
+      // adjusted one step back, etc.
+      range: null,
+
+      // The range for the X axis currently displayed in the graph, if zoomed.
+      zoomRange: null,
+    };
+
+    intersectionObserver.observe(this.container);
+    resizeObserver.observe(this.container);
   }
 
   async init() {
+    // Prevent execution of multiple initializations at the same time, due to
+    // the evaluation of the 'getMetric()' promise.
+    if (this.initializing) {
+      return;
+    }
+    this.initializing = true;
     try {
-      // Fetch metric, adjusting the effective step if necessary.
-      const metric = await this.getMetric();
-
-      // Calculate & store the range for the X axis. This may change during zoom
-      // events, and we need to know the original range to reset it.
-      this.range = this.getPlotlyXAxisRange(metric);
+      // Fetch metric (adjusting the effective step if necessary, overriding the
+      // selected aggregator if the metric is a bitmap, etc.) and immediately
+      // process it to prepare the graph for the first render.
+      this.processMetric(await this.getMetric());
 
       // Render the graph for the first time.
-      this.graph = this.renderGraph(metric);
+      this.graph.element = this.renderGraph();
 
       // Since the chart is being initialized, it is visible, so we can start
       // the refresh loop to keep the chart up-to-date.
       this.setupInterval();
+
+      // Initialize refresh flags:
+      //   - The 'last refresh' is needed because refresh loops are canceled
+      //     when the chart is not visible. When it becomes visible again,
+      //     the last refresh time is used to decide if the chart needs to be
+      //     refreshed right away or if it can wait for the next scheduled
+      //     refresh.
+      //   - The 'pending refresh' is needed because other properties (e.g, step,
+      //     aggregator, etc.) can change while the chart is not visible. When
+      //     the chart becomes visible again, it needs to be refreshed right
+      //     away to reflect these changes.
+      this.lastRefresh = helpers.dateToUnix(new Date());
+      this.pendingRefresh = false;
     } catch (error) {
       // On error, show some visual feedback to the user. Failing during the
       // initialization is a bit more complicated to recover automatically, so
       // we just show the error and let the user try to manually refresh the
       // chart.
       this.setError(`Failed to fetch samples of a metric: ${error}`);
+    } finally {
+      this.initializing = false;
     }
   }
 
@@ -65,53 +125,137 @@ class Chart {
   //
 
   async handleRefresh() {
-    // If the chart is not visible, there is no need to update it. Refresh loops
-    // are canceled when the chart is not visible, but this is a safety check.
-    // Also, manual refreshes would be pointless for invisible charts.
-    if (this.graph != null && this.visible) {
-      try {
-        // Fetch metric, adjusting the effective step if necessary.
-        const metric = await this.getMetric();
+    // Prevent execution of multiple refreshes at the same time, due to the
+    // evaluation of the 'getMetric()' promise.
+    if (this.refreshing) {
+      return;
+    }
+    this.refreshing = true;
+    try {
+      // Charts that are still pending initialization are not refreshed. This
+      // will be done when they are created for the first time.
+      if (this.graph.element != null) {
+        // Refreshing an already initialized chart that is not visible is a
+        // no-op, but a mark is used to trigger the pending refreshed as soon
+        // as it becomes visible again. Same for failed refreshes of visible
+        // charts.
+        this.pendingRefresh = true;
+        if (this.visible) {
+          try {
+            // Fetch metric (adjusting the effective step if necessary,
+            // overriding the selected aggregator if the metric is a bitmap,
+            // etc.) and immediately process it to prepare the graph for the
+            // update.
+            this.processMetric(await this.getMetric());
 
-        // Calculate & store the range for the X axis. This may change during
-        // zoom events, and we need to know the original range to reset it.
-        this.range = this.getPlotlyXAxisRange(metric);
+            // Update the graph with the new samples. For now we are just
+            // re-fetching and re-rendering the whole graph, but in the future
+            // we should be able to me more efficient. The server-side already
+            // supports this kind of incremental updates.
+            this.updateGraph(false);
 
-        // Update the graph with the new samples. For now we are just
-        // re-rendering the whole graph, but in the future we should be able to
-        // update it in a more efficient way. The server-side already supports
-        // this kind of incremental updates.
-        this.updateGraph(metric);
-
-        // All good, clear any previous error.
-        this.clearError();
-      } catch (error) {
-        // On error, show some visual feedback to the user. The next refresh,
-        // manual or automatic, will try to fetch the metric again and it will
-        // eventually succeed and clear the error.
-        this.setError(`Failed to fetch samples of a metric: ${error}`);
+            // All good, clear any previous error message and update the refresh
+            // flags.
+            this.clearError();
+            this.lastRefresh = helpers.dateToUnix(new Date());
+            this.pendingRefresh = false;
+          } catch (error) {
+            // On error, show some visual feedback to the user. The next refresh,
+            // manual or automatic, will try to fetch the metric again and it
+            // will eventually succeed and clear the error.
+            this.setError(`Failed to fetch samples of a metric: ${error}`);
+          }
+        }
       }
+    } finally {
+      this.refreshing = false;
+    }
+  }
+
+  handleVisibilityChange(visible) {
+    this.visible = visible;
+    if (this.visible) {
+      if (this.graph.element == null) {
+        this.init();
+      } else {
+        if (this.interval == null) {
+          this.setupInterval();
+        }
+        if (this.pendingRefresh ||
+            (this.refreshInterval > 0 && helpers.dateToUnix(new Date()) - this.lastRefresh > this.refreshInterval)) {
+          // Note that the debounced refresh handler is used here.
+          this.debouncedHandleRefresh();
+        }
+      }
+    } else {
+      this.stopInterval();
+    }
+  }
+
+  handleSizeChange() {
+    if (this.graph.element != null) {
+      // Note that the debounced refresh handler is used here.
+      this.debouncedHandleRefresh();
     }
   }
 
   handleGraphRelayout(event) {
+    // Extract the zoom range of the graph.
     if (event['xaxis.range[0]'] && event['xaxis.range[1]']) {
-      this.zoomRange = [
+      this.graph.zoomRange = [
         new Date(event['xaxis.range[0]']),
         new Date(event['xaxis.range[1]']),
       ];
     } else if (event['xaxis.range'] && Array.isArray(event['xaxis.range']) && event['xaxis.range'].length === 2) {
-      this.zoomRange = [
+      this.graph.zoomRange = [
         new Date(event['xaxis.range'][0]),
         new Date(event['xaxis.range'][1]),
       ];
     } else {
-      this.zoomRange = null;
+      this.graph.zoomRange = null;
     }
 
+    // If the zoom range goes beyond the original range, adjust it to the
+    // original range.
+    if (this.graph.zoomRange != null) {
+      if (helpers.dateToUnix(this.graph.zoomRange[0]) < helpers.dateToUnix(this.graph.range[0])) {
+        this.graph.zoomRange[0] = this.graph.range[0];
+      }
+      if (helpers.dateToUnix(this.graph.zoomRange[1]) > helpers.dateToUnix(this.graph.range[1])) {
+        this.graph.zoomRange[1] = this.graph.range[1];
+      }
+    }
+
+    // If the zoom range is too small, adjust it to a reasonable range.
+    if (this.graph.zoomRange != null) {
+      const difference = helpers.dateToUnix(this.graph.zoomRange[1]) - helpers.dateToUnix(this.graph.zoomRange[0]);
+      if (difference < this.graph.step) {
+        const center = helpers.dateToUnix(this.graph.zoomRange[0]) + Math.round(difference / 2);
+        this.graph.zoomRange = [
+          helpers.unixToDate(center - this.graph.step),
+          helpers.unixToDate(center + this.graph.step),
+        ];
+      }
+    }
+
+    // In the zoom range is identical to the original range, reset the zoom
+    // range so listeners can detect the no-zoom event.
+    if (this.graph.zoomRange != null) {
+      if (helpers.dateToUnix(this.graph.zoomRange[0]) === helpers.dateToUnix(this.graph.range[0]) &&
+          helpers.dateToUnix(this.graph.zoomRange[1]) === helpers.dateToUnix(this.graph.range[1])) {
+        this.graph.zoomRange = null;
+      }
+    }
+
+    // Force a re-render of the graph. The zoom range is already applied, but we
+    // might want to adjust other properties of the graph (e.g., the data mode),
+    // or the adjusted zoom range itself.
+    this.updateGraph(true);
+
+    // Inform listeners about the zoom event.
     this.notifyEventListeners('zoom', {
       target: this,
-      range: this.zoomRange,
+      range: this.graph.zoomRange,
     });
   }
 
@@ -142,20 +286,8 @@ class Chart {
   // Public API.
   //
 
-  setVisible(visible) {
-    this.visible = visible;
-    if (this.visible) {
-      if (this.graph == null) {
-        this.init();
-      } else if (this.interval == null) {
-        this.setupInterval();
-      }
-    } else {
-      this.stopInterval();
-    }
-  }
-
   redraw(filter, verbosity, columns) {
+    // Hide the chart if the metric name does not match the filter & verbosity.
     let hidden = verbosity === 'normal' && this.metric.debug;
     if (!hidden) {
       const terms = filter.split(/\s+/).filter(term => term.length > 0);
@@ -165,6 +297,7 @@ class Chart {
     }
     this.container.classList.toggle('d-none', hidden);
 
+    // Rearrange CSS classes to adjust to the number of columns.
     this.container.classList.forEach(className => {
       if (className.startsWith('col-')) {
         this.container.classList.remove(className);
@@ -172,57 +305,66 @@ class Chart {
     });
     this.container.classList.add(`col-${12 / columns}`);
 
-    if (this.graph != null) {
-      Plotly.Plots.resize(this.graph);
-    }
+    // No need to call 'handleRefresh()' here: visibility changes are managed by
+    // 'handleVisibilityChange()', and size changes are managed by
+    // 'handleSizeChange()'.
   }
 
   setRefreshInterval(interval) {
     this.refreshInterval = interval;
-    this.setupInterval();
+    if (this.visible && this.graph.element != null) {
+      this.setupInterval();
+    }
   }
 
   refresh() {
-    this.setupInterval();
-    this.handleRefresh();
+    if (this.graph.element != null) {
+      if (this.visible) {
+        this.setupInterval();
+      }
+      this.handleRefresh();
+    }
   }
 
   setAggregator(aggregator) {
     this.aggregator = aggregator;
-    this.setupInterval();
-    this.handleRefresh();
+    if (this.graph.element != null) {
+      if (this.visible) {
+        this.setupInterval();
+      }
+      this.handleRefresh();
+    }
   }
 
   setStep(step) {
     this.step = step;
-    this.setupInterval();
-    this.handleRefresh();
+    if (this.graph.element != null) {
+      if (this.visible) {
+        this.setupInterval();
+      }
+      this.handleRefresh();
+    }
   }
 
   setZoomRange(range) {
-    // Ignore if the range is the same as the current one.
-    if (this.zoomRange === range) {
-      return;
-    }
-    this.zoomRange = range;
+    this.graph.zoomRange = range;
 
-    // Adjust range in the X axis. It's important to apply the adjustment to
-    // every initialized chart, even if they are not visible.
-    if (this.graph != null) {
-      Plotly.update(this.graph, [], {
-        xaxis: {
-          range: this.zoomRange != null ? this.zoomRange : this.range,
-        },
-      });
+    // If the chart has already been initialized (whether visible or not),
+    // update it to reflect the new range, the potential new data mode, etc.
+    // Beware 'handleRefresh()' is not called here, as we are not fetching new
+    // samples, just updating the graph with the current data.
+    if (this.graph.element != null) {
+      this.updateGraph(true);
     }
   }
 
   destroy() {
-    observer.unobserve(this.container);
+    intersectionObserver.unobserve(this.container);
+    resizeObserver.unobserve(this.container);
     this.stopInterval();
     this.clearError();
-    if (this.graph != null) {
-      Plotly.purge(this.graph);
+    if (this.graph.element != null) {
+      Plotly.purge(this.graph.element);
     }
   }
 
@@ -232,13 +374,9 @@ class Chart {
 
   setupInterval() {
     this.stopInterval();
-
-    if (this.visible) {
-      if (this.graph != null && this.refreshInterval > 0) {
-        this.interval = setInterval(this.handleRefresh.bind(this), this.refreshInterval * 1000);
-      } else if (this.graph == null) {
-        this.init();
-      }
+    if (this.refreshInterval > 0) {
+      this.interval = setInterval(
+        this.handleRefresh.bind(this), this.refreshInterval * 1000);
     }
   }
 
@@ -317,6 +455,12 @@ class Chart {
     const containerWidth = this.container.clientWidth;
     const maxSamples = Math.floor(0.9 * containerWidth);
 
+    // Stop here if the maximum number of samples cannot be calculated (most
+    // likely because the container is not visible due to a race condition).
+    if (maxSamples <= 0) {
+      throw new Error('Failed to estimate the optimal step');
+    }
+
     // If the number of samples required to cover the whole range is less than
     // the number of samples that would fit in the graph, we can use the current
     // step. Otherwise, we need to calculate a new step that would fit the
@@ -331,19 +475,60 @@ class Chart {
     return Math.ceil(samples / maxSamples) * this.step;
   }
 
-  renderGraph(metric) {
+  processMetric(metric) {
+    // Prepare X & Y data for Plotly.
+    this.graph.x = [];
+    this.graph.y = [];
+    metric.samples.forEach(sample => {
+      this.graph.x.push(sample[0]);
+      // Bitmap metrics are returned as an hex string. For now, we represent the
+      // number of bits set to 1 in the bitmap as the Y value. This will be
+      // improved in the future using a different visualization for bitmap
+      // metrics.
+      this.graph.y.push(
+        this.metric.flag === 'b' ?
+          BigInt(`0x${sample[1]}`).toString(2).split('').filter(bit => bit === '1').length :
+          sample[1]);
+    });
+
+    // Store the step (already adjusted to be optimal for the space available)
+    // as returned by the storage.
+    this.graph.step = metric.step;
+
+    // Calculate & store the range for the X axis. This may change during
+    // zoom events, and we need to know the original range to reset it.
+    this.graph.range = [
+      metric.from,
+      new Date(metric.to.getTime() - metric.step * 1000),
+    ];
+
+    // If the effective step is different from the selected step, show in the
+    // card the factor by which the selected step has been adjusted.
+    const stepFactor = this.container.querySelector('.card .step-factor');
+    if (this.graph.step !== this.step) {
+      const factor = Math.round(this.graph.step / this.step);
+      stepFactor.innerHTML =
+        `<i class="fa-solid fa-arrows-left-right-to-line"></i> ${factor}x`;
+    } else {
+      stepFactor.innerHTML = '';
+    }
+  }
+
+  renderGraph() {
+    // Decide range to be used in the X axis.
+    const range = this.graph.zoomRange != null ? this.graph.zoomRange : this.graph.range;
+
     // Prepare data for Plotly.
-    const [x, y] = this.getPlotlyDataXY(metric);
     const data = [
       {
-        x: x,
-        y: y,
+        x: this.graph.x,
+        y: this.graph.y,
         type: 'scatter',
-        mode: this.estimatePlotlyDataMode(metric),
+        mode: this.estimatePlotlyDataMode(...range, this.graph.step),
         marker: { size: 4 },
         hovertemplate: '<b>X:</b> %{x|%Y-%m-%d %H:%M:%S}<br><b>Y:</b> %{y:,.1f}<extra></extra>',
         connectgaps: false,
-        line: { shape: 'linear', width: 1 },
+        line: { shape: 'linear', width: 2 },
       }
     ];
 
@@ -359,7 +544,7 @@ class Chart {
       xaxis: {
         fixedrange: false,
         griddash: 'dash',
-        range: this.zoomRange != null ? this.zoomRange : this.range,
+        range: Array.from(range), // Beware the array needs to be cloned.
         autorange: false,
       },
       yaxis: {
@@ -398,58 +583,46 @@ class Chart {
         filename: `${varnishmon.storage.hostname} - ${this.metric.name}`,
         format: 'png',
       },
+      scrollZoom: false,
     };
 
     // Render the graph.
-    const graph = this.container.querySelector('.graph');
-    Plotly.newPlot(graph, data, layout, config);
+    const graphElement = this.container.querySelector('.graph');
+    Plotly.newPlot(graphElement, data, layout, config);
 
     // Handle Plotly events.
-    graph.on('plotly_relayout', this.handleGraphRelayout.bind(this));
+    graphElement.on('plotly_relayout', this.handleGraphRelayout.bind(this));
 
     // Done!
-    return graph;
+    return graphElement;
   }
 
-  updateGraph(metric) {
+  updateGraph(sameData) {
+    // Decide range to be used in the X axis.
+    const range = this.graph.zoomRange != null ? this.graph.zoomRange : this.graph.range;
+
     // Prepare data for Plotly.
-    const [x, y] = this.getPlotlyDataXY(metric);
     const data = {
-      x: [x],
-      y: [y],
-      mode: this.estimatePlotlyDataMode(metric),
+      mode: this.estimatePlotlyDataMode(...range, this.graph.step),
     };
+    if (!sameData) {
+      data.x = [this.graph.x];
+      data.y = [this.graph.y];
+    }
 
     // Prepare layout for Plotly.
     const layout = {
       xaxis: {
-        range: this.zoomRange != null ? this.zoomRange : this.range,
+        range: Array.from(range), // Beware the array needs to be cloned.
       },
     };
 
     // Update the graph!
-    Plotly.update(this.graph, data, layout);
+    Plotly.update(this.graph.element, data, layout);
   }
 
-  getPlotlyDataXY(metric) {
-    const x = [];
-    const y = [];
-    metric.samples.forEach(sample => {
-      x.push(sample[0]);
-      // Bitmap metrics are returned as an hex string. For now, we represent the
-      // number of bits set to 1 in the bitmap as the Y value. This will be
-      // improved in the future using a different visualization for bitmap
-      // metrics.
-      y.push(
-        this.metric.flag === 'b' ?
-          BigInt(`0x${sample[1]}`).toString(2).split('').filter(bit => bit === '1').length :
-          sample[1]);
-    });
-    return [x, y];
-  }
-
-  estimatePlotlyDataMode(metric) {
-    const samples = (helpers.dateToUnix(metric.to) - helpers.dateToUnix(metric.from)) / metric.step;
+  estimatePlotlyDataMode(from, to, step) {
+    const samples = (helpers.dateToUnix(to) - helpers.dateToUnix(from)) / step;
     const containerWidth = this.container.clientWidth;
     const minSpacing = 6;
     const maxSamples = Math.floor(0.9 * containerWidth / minSpacing);
@@ -457,10 +630,6 @@ class Chart {
       return 'lines';
     }
     return 'lines+markers';
-  }
-
-  getPlotlyXAxisRange(metric) {
-    return [metric.from, new Date(metric.to.getTime() - metric.step * 1000)];
   }
 }